requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
waitress==2.1.2
email-validator==2.0.0